+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## 1.0.6
+
+Enhancements:
+
+ -- added profiler utils
+
+## 1.0.5
+
+Enhancements:
+
+ -- added HTML only email option
+
 ## 1.0.4
 
 BugFix: